requirements.txt
mkdocs-material==9.*
mkdocs-material-extensions==1.*
python-markdown-math==0.8
mkdocs-macros-plugin==1.*
